requirements.txt
neo4j==5.27.0
streamlit==1.41.0
anthropic==0.42.0
sentence-transformers==3.3.1
python-dotenv==1.0.1
pyvis==0.3.2
streamlit-agraph==0.0.45
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
